Refactored pages-component structure
