+++ src/pages-components/ProductsPage/ui/ProductsPage/ProductsFilters/CollectionFilters/CollectionFilters.module.scss
