+++ src/pages-components/ProductsPage/ui/ProductsPage/ProductsFilters/FilterItem/FilterItem.module.scss
