+++ src/pages-components/ProductsPage/ui/ProductsPage/ProductsFilters/PriceFilters/PriceFilters.module.scss
