+++ src/pages-components/ProductsPage/ui/ProductsPage/ProductsFilters/ProductsFilters.tsx
@@ -1,13 +1,13 @@
 import { FC } from 'react';
 import { classNames } from 'shared/lib/classNames/classNames';
 import cls from './ProductsFilters.module.scss';
-import CategoryFilters from '../CategoryFilters/CategoryFilters';
-import PriceFilters from '../PriceFilters/PriceFilters';
-import SizeFilters from '../SizeFilters/SizeFilters';
-import CollectionFilters from '../CollectionFilters/CollectionFilters';
-import BrandFilters from '../BrandFilters/BrandFilters';
-import StyleFilters from '../StyleFilters/StyleFilters';
-import SeasonFilters from '../SeasonFilters/SeasonFilters';
+import CategoryFilters from './CategoryFilters/CategoryFilters';
+import PriceFilters from './PriceFilters/PriceFilters';
+import SizeFilters from './SizeFilters/SizeFilters';
+import CollectionFilters from './CollectionFilters/CollectionFilters';
+import BrandFilters from './BrandFilters/BrandFilters';
+import StyleFilters from './StyleFilters/StyleFilters';
+import SeasonFilters from './SeasonFilters/SeasonFilters';
 
 interface ProductsFiltersProps {
   className?: string;
