+++ src/pages-components/ProductsPage/ui/ProductsPage/ProductsFilters/StyleFilters/StyleFilters.module.scss
