+++ src/pages-components/RegisterPage/ui/RegisterPage/RegisterTopControls/RegisterTopControls.module.scss
