+++ src/pages-components/RegisterPage/ui/RegisterPage/RegisterTopControls/RegisterTopControls.tsx
