+++ src/widgets/ProductItem/ui/ProductItem/ProductItem.module.scss
@@ -2,6 +2,7 @@
   height: 48.09vh;
   height: 450px;
   width: 300px;
+  width: 100%;
   cursor: pointer;
   transition: transform 0.15s ease-in-out;
   &:hover {
